refactor folder models, usecase
