+++ app/src/main/res/layout/list_item.xml
@@ -5,7 +5,7 @@
 
         <variable
             name="item"
-            type="com.example.native_new.android.androidbasejava.data.db.entities.Books" />
+            type="com.example.native_new.android.androidbasejava.data.models.Books" />
     </data>
 
     <androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
